+++ core/common/src/main/kotlin/com/willfp/libreforge/integrations/xiaomomiplugins/customcrops/impl/FilterFertilizerType.kt
@@ -1,19 +1,19 @@
-package com.willfp.libreforge.integrations.customcrops.impl
+package com.willfp.libreforge.integrations.xiaomomiplugins.customcrops.impl
 
 import com.willfp.eco.core.config.interfaces.Config
 import com.willfp.libreforge.NoCompileData
 import com.willfp.libreforge.filters.Filter
 import com.willfp.libreforge.triggers.TriggerData
-import net.momirealms.customcrops.api.event.CropBreakEvent
+import net.momirealms.customcrops.api.event.FertilizerUseEvent
 
-object FilterCustomCropStage : Filter<NoCompileData, Collection<String>>("custom_crop_stage") {
+object FilterFertilizerType : Filter<NoCompileData, Collection<String>>("fertilizer_type") {
     override fun getValue(config: Config, data: TriggerData?, key: String): Collection<String> {
         return config.getStrings(key)
     }
 
     override fun isMet(data: TriggerData, value: Collection<String>, compileData: NoCompileData): Boolean {
-        val event = data.event as? CropBreakEvent ?: return true
+        val event = data.event as? FertilizerUseEvent ?: return true
 
-        return value.contains(event.cropStageItemID())
+        return value.contains(event.fertilizer().id())
     }
 }
