+++ core/common/src/main/kotlin/com/willfp/libreforge/integrations/xiaomomiplugins/customcrops/impl/TriggerBonemealCrop.kt
@@ -1,13 +1,13 @@
-package com.willfp.libreforge.integrations.customcrops.impl
+package com.willfp.libreforge.integrations.xiaomomiplugins.customcrops.impl
 
 import com.willfp.libreforge.toDispatcher
 import com.willfp.libreforge.triggers.Trigger
 import com.willfp.libreforge.triggers.TriggerData
 import com.willfp.libreforge.triggers.TriggerParameter
-import net.momirealms.customcrops.api.event.CropPlantEvent
+import net.momirealms.customcrops.api.event.BoneMealUseEvent
 import org.bukkit.event.EventHandler
 
-object TriggerPlantCustomCrop : Trigger("plant_custom_crop") {
+object TriggerBonemealCrop : Trigger("bonemeal_crop") {
     override val parameters = setOf(
         TriggerParameter.PLAYER,
         TriggerParameter.EVENT,
@@ -16,7 +16,7 @@ object TriggerPlantCustomCrop : Trigger("plant_custom_crop") {
     )
 
     @EventHandler(ignoreCancelled = true)
-    fun handle(event: CropPlantEvent) {
+    fun handle(event: BoneMealUseEvent) {
         val player = event.player ?: return
         val location = event.location() ?: return
         val crop = event.cropConfig().id() ?: return
