+++ core/common/src/main/kotlin/com/willfp/libreforge/integrations/xiaomomiplugins/customcrops/impl/TriggerHarvestCustomCrop.kt
@@ -1,13 +1,14 @@
-package com.willfp.libreforge.integrations.customcrops.impl
+package com.willfp.libreforge.integrations.xiaomomiplugins.customcrops.impl
 
 import com.willfp.libreforge.toDispatcher
 import com.willfp.libreforge.triggers.Trigger
 import com.willfp.libreforge.triggers.TriggerData
 import com.willfp.libreforge.triggers.TriggerParameter
-import net.momirealms.customcrops.api.event.BoneMealUseEvent
+import net.momirealms.customcrops.api.event.CropBreakEvent
+import org.bukkit.entity.Player
 import org.bukkit.event.EventHandler
 
-object TriggerBonemealCrop : Trigger("bonemeal_crop") {
+object TriggerHarvestCustomCrop : Trigger("harvest_custom_crop") {
     override val parameters = setOf(
         TriggerParameter.PLAYER,
         TriggerParameter.EVENT,
@@ -16,8 +17,8 @@ object TriggerBonemealCrop : Trigger("bonemeal_crop") {
     )
 
     @EventHandler(ignoreCancelled = true)
-    fun handle(event: BoneMealUseEvent) {
-        val player = event.player ?: return
+    fun handle(event: CropBreakEvent) {
+        val player = event.entityBreaker() as? Player ?: return
         val location = event.location() ?: return
         val crop = event.cropConfig().id() ?: return
