+++ core/common/src/main/kotlin/com/willfp/libreforge/integrations/xiaomomiplugins/customcrops/impl/TriggerUseFertilizer.kt
@@ -1,32 +1,33 @@
-package com.willfp.libreforge.integrations.customfishing.impl
+package com.willfp.libreforge.integrations.xiaomomiplugins.customcrops.impl
 
 import com.willfp.libreforge.toDispatcher
 import com.willfp.libreforge.triggers.Trigger
 import com.willfp.libreforge.triggers.TriggerData
 import com.willfp.libreforge.triggers.TriggerParameter
-import net.momirealms.customfishing.api.event.FishingResultEvent
+import net.momirealms.customcrops.api.event.FertilizerUseEvent
 import org.bukkit.event.EventHandler
 
-object TriggerCatchFishFail : Trigger("catch_fish_fail") {
+object TriggerUseFertilizer : Trigger("use_fertilizer") {
     override val parameters = setOf(
         TriggerParameter.PLAYER,
         TriggerParameter.EVENT,
-        TriggerParameter.TEXT
+        TriggerParameter.TEXT,
+        TriggerParameter.LOCATION
     )
 
     @EventHandler(ignoreCancelled = true)
-    fun handle(event: FishingResultEvent) {
+    fun handle(event: FertilizerUseEvent) {
         val player = event.player ?: return
-        val result = event.result ?: return
-
-        if (result != FishingResultEvent.Result.FAILURE) return
+        val location = event.location() ?: return
+        val fertilizer = event.fertilizer().id() ?: return
 
         this.dispatch(
             player.toDispatcher(),
             TriggerData(
                 player = player,
                 event = event,
-                text = result.name
+                text = fertilizer,
+                location = location
             )
         )
     }
