+++ core/common/src/main/kotlin/com/willfp/libreforge/integrations/xiaomomiplugins/customcrops/impl/TriggerUseWateringCan.kt
@@ -1,13 +1,14 @@
-package com.willfp.libreforge.integrations.customcrops.impl
+package com.willfp.libreforge.integrations.xiaomomiplugins.customcrops.impl
 
 import com.willfp.libreforge.toDispatcher
 import com.willfp.libreforge.triggers.Trigger
 import com.willfp.libreforge.triggers.TriggerData
 import com.willfp.libreforge.triggers.TriggerParameter
-import net.momirealms.customcrops.api.event.FertilizerUseEvent
+import net.momirealms.customcrops.api.event.WateringCanWaterPotEvent
+import net.momirealms.customcrops.api.event.WateringCanWaterSprinklerEvent
 import org.bukkit.event.EventHandler
 
-object TriggerUseFertilizer : Trigger("use_fertilizer") {
+object TriggerUseWateringCan : Trigger("use_watering_can") {
     override val parameters = setOf(
         TriggerParameter.PLAYER,
         TriggerParameter.EVENT,
@@ -16,18 +17,29 @@ object TriggerUseFertilizer : Trigger("use_fertilizer") {
     )
 
     @EventHandler(ignoreCancelled = true)
-    fun handle(event: FertilizerUseEvent) {
+    fun handle(event: WateringCanWaterSprinklerEvent) {
         val player = event.player ?: return
         val location = event.location() ?: return
-        val fertilizer = event.fertilizer().id() ?: return
 
         this.dispatch(
             player.toDispatcher(),
             TriggerData(
                 player = player,
                 event = event,
-                text = fertilizer,
                 location = location
+            )
+        )
+    }
+
+    @EventHandler(ignoreCancelled = true)
+    fun handle(event: WateringCanWaterPotEvent) {
+        val player = event.player ?: return
+
+        this.dispatch(
+            player.toDispatcher(),
+            TriggerData(
+                player = player,
+                event = event,
             )
         )
     }
