+++ core/common/src/main/kotlin/com/willfp/libreforge/integrations/xiaomomiplugins/customfishing/CustomFishingIntegration.kt
@@ -1,11 +1,11 @@
-package com.willfp.libreforge.integrations.customcrops
+package com.willfp.libreforge.integrations.xiaomomiplugins.customfishing
 
 import com.willfp.eco.core.EcoPlugin
 import com.willfp.libreforge.filters.Filters
 import com.willfp.libreforge.integrations.LoadableIntegration
-import com.willfp.libreforge.integrations.customfishing.impl.FilterCustomFishType
-import com.willfp.libreforge.integrations.customfishing.impl.TriggerCatchFish
-import com.willfp.libreforge.integrations.customfishing.impl.TriggerCatchFishFail
+import com.willfp.libreforge.integrations.xiaomomiplugins.customfishing.impl.FilterCustomFishType
+import com.willfp.libreforge.integrations.xiaomomiplugins.customfishing.impl.TriggerCatchFish
+import com.willfp.libreforge.integrations.xiaomomiplugins.customfishing.impl.TriggerCatchFishFail
 import com.willfp.libreforge.triggers.Triggers
 
 object CustomFishingIntegration : LoadableIntegration {
