+++ core/common/src/main/kotlin/com/willfp/libreforge/integrations/xiaomomiplugins/customfishing/impl/TriggerCatchFishFail.kt
@@ -1,34 +1,32 @@
-package com.willfp.libreforge.integrations.customcrops.impl
+package com.willfp.libreforge.integrations.xiaomomiplugins.customfishing.impl
 
 import com.willfp.libreforge.toDispatcher
 import com.willfp.libreforge.triggers.Trigger
 import com.willfp.libreforge.triggers.TriggerData
 import com.willfp.libreforge.triggers.TriggerParameter
-import net.momirealms.customcrops.api.event.CropBreakEvent
-import org.bukkit.entity.Player
+import net.momirealms.customfishing.api.event.FishingResultEvent
 import org.bukkit.event.EventHandler
 
-object TriggerHarvestCustomCrop : Trigger("harvest_custom_crop") {
+object TriggerCatchFishFail : Trigger("catch_fish_fail") {
     override val parameters = setOf(
         TriggerParameter.PLAYER,
         TriggerParameter.EVENT,
-        TriggerParameter.TEXT,
-        TriggerParameter.LOCATION
+        TriggerParameter.TEXT
     )
 
     @EventHandler(ignoreCancelled = true)
-    fun handle(event: CropBreakEvent) {
-        val player = event.entityBreaker() as? Player ?: return
-        val location = event.location() ?: return
-        val crop = event.cropConfig().id() ?: return
+    fun handle(event: FishingResultEvent) {
+        val player = event.player ?: return
+        val result = event.result ?: return
+
+        if (result != FishingResultEvent.Result.FAILURE) return
 
         this.dispatch(
             player.toDispatcher(),
             TriggerData(
                 player = player,
                 event = event,
-                text = crop,
-                location = location
+                text = result.name
             )
         )
     }
